+++ lib/main.dart
@@ -1,8 +1,10 @@
 import 'package:dynamic_color/dynamic_color.dart';
+import 'package:feedback/feedback.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:provider/provider.dart';
 import 'package:tarefas/auth/auth.dart';
 import 'package:tarefas/firebase_options.dart';
@@ -19,7 +21,18 @@ main() async {
 
   FlutterError.onError = FirebaseCrashlytics.instance.recordFlutterFatalError;
 
-  runApp(const MyApp());
+  runApp(BetterFeedback(
+    theme: FeedbackThemeData.light(),
+    darkTheme: FeedbackThemeData.dark(),
+    localizationsDelegates: [
+      GlobalMaterialLocalizations.delegate,
+      GlobalCupertinoLocalizations.delegate,
+      GlobalWidgetsLocalizations.delegate,
+      GlobalFeedbackLocalizationsDelegate(),
+    ],
+    localeOverride: const Locale('pt'),
+    child: const MyApp(),
+  ));
 }
 
 class MyApp extends StatelessWidget {
@@ -33,48 +46,56 @@ class MyApp extends StatelessWidget {
       return ChangeNotifierProvider(
         create: (_) => ThemeModel(),
         child: Consumer<ThemeModel>(builder: (_, theme, __) {
-          return MaterialApp(
-              theme: ThemeData(
-                brightness: Brightness.light,
-                colorScheme: lightColorScheme?.copyWith(
-                  primary: theme.isDarkMode ? Colors.black : Colors.black,
+          return DynamicColorBuilder(
+              builder: (lightColorScheme, darkColorScheme) {
+            if (!theme.isDynamicColorsEnabled) {
+              lightColorScheme = null;
+              darkColorScheme = null;
+            }
+
+            return MaterialApp(
+                theme: ThemeData(
+                  brightness: Brightness.light,
+                  colorScheme: lightColorScheme?.copyWith(
+                    primary: theme.isDarkMode ? Colors.black : Colors.black,
+                  ),
+                  useMaterial3: true,
+                  textTheme: Typography().black.apply(),
                 ),
-                useMaterial3: true,
-                textTheme: Typography().black.apply(),
-              ),
-              darkTheme: ThemeData(
-                brightness: Brightness.dark,
-                colorScheme: darkColorScheme?.copyWith(
-                  primary: theme.isDarkMode ? Colors.white : Colors.black,
+                darkTheme: ThemeData(
+                  brightness: Brightness.dark,
+                  colorScheme: darkColorScheme?.copyWith(
+                    primary: theme.isDarkMode ? Colors.white : Colors.black,
+                  ),
+                  useMaterial3: true,
+                  textTheme: Typography().white.apply(),
                 ),
-                useMaterial3: true,
-                textTheme: Typography().white.apply(),
-              ),
-              themeMode: theme.isDarkMode ? ThemeMode.dark : ThemeMode.light,
-              debugShowCheckedModeBanner: false,
-              home: FutureBuilder(
-                future: authService.currentUser(),
-                builder: (context, AsyncSnapshot<User?> snapshot) {
-                  if (snapshot.connectionState == ConnectionState.done) {
-                    if (snapshot.hasData) {
-                      return const BottomNavigationContainer();
+                themeMode: theme.isDarkMode ? ThemeMode.dark : ThemeMode.light,
+                debugShowCheckedModeBanner: false,
+                home: FutureBuilder(
+                  future: authService.currentUser(),
+                  builder: (context, AsyncSnapshot<User?> snapshot) {
+                    if (snapshot.connectionState == ConnectionState.done) {
+                      if (snapshot.hasData) {
+                        return const BottomNavigationContainer();
+                      } else {
+                        return LoginScreen(
+                          authService: authService,
+                        );
+                      }
                     } else {
-                      return LoginScreen(
-                        authService: authService,
+                      return const Scaffold(
+                        body: Center(
+                          child: CircularProgressIndicator.adaptive(),
+                        ),
                       );
                     }
-                  } else {
-                    return const Scaffold(
-                      body: Center(
-                        child: CircularProgressIndicator.adaptive(),
-                      ),
-                    );
-                  }
-                },
-              ),
-              routes: {
-                '/login': (context) => LoginScreen(authService: authService),
-              });
+                  },
+                ),
+                routes: {
+                  '/login': (context) => LoginScreen(authService: authService),
+                });
+          });
         }),
       );
     });